remove handlers to import, cleans up main index majorly
removes tons of imports, now each has a function in the client folder
